+++ doc/changelog.rst
@@ -7,6 +7,7 @@ What's New
 * Removes unneeded ``--no-strong-robots`` (superseded with ``--no-strong-redirects``.)
 * Fixes ``--no-phantomjs-snapshot`` option not respected.
 * More link extraction on HTML pages with elements with ``onclick``, ``onkeyX``, ``onmouseX``, and ``data-`` attributes.
+* Adds web-based debugging console with ``--debug-console-port``.
 
 
 0.33.2 (2014-04-29)